+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+#### Builtin
+
+- add ``UnicodeData`` type
+    - string representation of object is Unicode version string (equivalent to ``shctl info`` output)
+    - add ``category`` method for query Unicode general category of code point (single utf8 char)
+
 ## [0.40.0] - 2025-12-30
 
 ### Added